Move labels/labels.go to labels.go
