+++ pkg/controller/apmserver/controller.go
@@ -15,7 +15,6 @@ import (
 	"go.elastic.co/apm"
 
 	apmv1 "github.com/elastic/cloud-on-k8s/pkg/apis/apm/v1"
-	"github.com/elastic/cloud-on-k8s/pkg/controller/apmserver/labels"
 	apmname "github.com/elastic/cloud-on-k8s/pkg/controller/apmserver/name"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/association"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common"
@@ -211,7 +210,7 @@ func (r *ReconcileApmServer) Reconcile(request reconcile.Request) (reconcile.Res
 }
 
 func (r *ReconcileApmServer) isCompatible(ctx context.Context, as *apmv1.ApmServer) (bool, error) {
-	selector := map[string]string{labels.ApmServerNameLabelName: as.Name}
+	selector := map[string]string{ApmServerNameLabelName: as.Name}
 	compat, err := annotation.ReconcileCompatibility(ctx, r.Client, as, selector, r.OperatorInfo.BuildInfo.Version)
 	if err != nil {
 		k8s.EmitErrorEvent(r.recorder, err, as, events.EventCompatCheckError, "Error during compatibility check: %v", err)
@@ -237,7 +236,7 @@ func (r *ReconcileApmServer) doReconcile(ctx context.Context, request reconcile.
 		Object:                as,
 		TLSOptions:            as.Spec.HTTP.TLS,
 		Namer:                 apmname.APMNamer,
-		Labels:                labels.NewLabels(as.Name),
+		Labels:                NewLabels(as.Name),
 		Services:              []corev1.Service{*svc},
 		CACertRotation:        r.CACertRotation,
 		CertRotation:          r.CertRotation,
@@ -297,7 +296,7 @@ func reconcileApmServerToken(c k8s.Client, as *apmv1.ApmServer) (corev1.Secret, 
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: as.Namespace,
 			Name:      apmname.SecretToken(as.Name),
-			Labels:    common.AddCredentialsLabel(labels.NewLabels(as.Name)),
+			Labels:    common.AddCredentialsLabel(NewLabels(as.Name)),
 		},
 		Data: make(map[string][]byte),
 	}
@@ -322,7 +321,7 @@ func (r *ReconcileApmServer) deploymentParams(
 ) (deployment.Params, error) {
 
 	podSpec := newPodSpec(as, params)
-	podLabels := labels.NewLabels(as.Name)
+	podLabels := NewLabels(as.Name)
 
 	// Build a checksum of the configuration, add it to the pod labels so a change triggers a rolling update
 	configChecksum := sha256.New224()
@@ -397,8 +396,8 @@ func (r *ReconcileApmServer) deploymentParams(
 		Name:            apmname.Deployment(as.Name),
 		Namespace:       as.Namespace,
 		Replicas:        as.Spec.Count,
-		Selector:        labels.NewLabels(as.Name),
-		Labels:          labels.NewLabels(as.Name),
+		Selector:        NewLabels(as.Name),
+		Labels:          NewLabels(as.Name),
 		PodTemplateSpec: podSpec,
 		Strategy:        appsv1.RollingUpdateDeploymentStrategyType,
 	}, nil
@@ -425,7 +424,7 @@ func (r *ReconcileApmServer) reconcileApmServerDeployment(
 		r,
 		as,
 		apmname.APMNamer,
-		labels.NewLabels(as.Name),
+		NewLabels(as.Name),
 		initContainerParameters,
 	)
 	if err != nil {
@@ -486,7 +485,7 @@ func NewService(as apmv1.ApmServer) *corev1.Service {
 	svc.ObjectMeta.Namespace = as.Namespace
 	svc.ObjectMeta.Name = apmname.HTTPService(as.Name)
 
-	labels := labels.NewLabels(as.Name)
+	labels := NewLabels(as.Name)
 	ports := []corev1.ServicePort{
 		{
 			Name:     as.Spec.HTTP.Protocol(),
