+++ pkg/controller/kibana/driver.go
@@ -18,7 +18,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
-	"github.com/elastic/cloud-on-k8s/pkg/controller/apmserver/labels"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/apmserver"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/association"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/certificates"
@@ -117,7 +117,7 @@ func (d *driver) Reconcile(
 		Object:                kb,
 		TLSOptions:            kb.Spec.HTTP.TLS,
 		Namer:                 Namer,
-		Labels:                labels.NewLabels(kb.Name),
+		Labels:                apmserver.NewLabels(kb.Name),
 		Services:              []corev1.Service{*svc},
 		CACertRotation:        params.CACertRotation,
 		CertRotation:          params.CertRotation,
